explicit casting of items to the 'right' size for use

+++ mgmt/cluster/ClusterCom.cc
@@ -2460,7 +2460,7 @@ checkBackDoor(int req_fd, char *message)
 // TODO: Was #if defined(solaris) && (!defined(_FILE_OFFSET_BITS) || _FILE_OFFSET_BITS != 64)
 #if defined(solaris)
     snprintf(reply, sizeof(reply), "\twatched_process_fd: %d  watched_process_pid: %ld\n",
-             lmgmt->watched_process_fd, lmgmt->watched_process_pid);
+             lmgmt->watched_process_fd, (long int)lmgmt->watched_process_pid);
 #else
     snprintf(reply, sizeof(reply), "\twatched_process_fd: %d  watched_process_pid: %d\n",
              lmgmt->watched_process_fd, lmgmt->watched_process_pid);
